Complied connect var data in GL to D3D for the wsPosition Made the use of ints for the cubemapIdx in the shaders consistent

+++ Templates/BaseGame/game/core/rendering/shaders/gl/lighting.glsl
@@ -462,7 +462,7 @@ vec4 computeForwardProbes(Surface surface,
       float contrib = contribution[i];
       if (contrib > 0.0f)
       {
-         float cubemapIdx = int(inProbeConfigData[i].a);
+         int cubemapIdx = int(inProbeConfigData[i].a);
          vec3 dir = boxProject(surface.P, surface.R, inWorldToObjArray[i], inRefScaleArray[i].xyz, inRefPosArray[i].xyz);
 
          irradiance += textureLod(irradianceCubemapAR, vec4(dir, cubemapIdx), 0).xyz * contrib;
